scripts/simulate-incidents/requirements.txt
azure-identity>=1.15.0
azure-cosmos>=4.7.0
requests>=2.31.0
